Add item-details components folder, itemdetails page and [id] file for dynamic routing
