+++ src/components/item-details/detail-card.tsx
