+++ src/components/item-details/detail-images-carousel.tsx
