+++ src/components/item-details/detail-map.module.css
