+++ src/components/item-details/detail-map.tsx
